backend/foodgram/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
django-filter==22.1
djoser==2.1.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
django-cors-headers==3.13.0
django-enumchoicefield==3.0.1
gunicorn==20.0.4
